backend/requirements.txt
Django
djangorestframework
django-cors-headers
langchain
langchain-ollama
langchain-chroma
pandas
requests
beautifulsoup4
chromadb
lxml